Changed int to size_t, shameless lint(1) footrub.
git-svn-id: file:///Users/kling/src/protoman/svn/trunk/ytalk@72 8120b729-f3fe-0310-9465-b9035d29392c

+++ src/fd.c
@@ -99,7 +99,7 @@ int
 full_read(fd, buf, len)
   int fd;
   register char *buf;
-  register int len;
+  register size_t len;
 {
     register int rc;
